content/html/content/test/test_bug599975.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=599975
-->
<head>
  <title>Test for Bug 599975</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=599975">Mozilla Bug 599975</a>
<p id="display"></p>
<div id="content" style="display: none">
<img id=image>
</div>
<pre id="test">
<script type="application/javascript">
/** Test for Bug 599975 **/
SimpleTest.waitForExplicitFinish();
var img = document.getElementById("image");
img.onerror = function() {
  ok(true, "onerror fired");
  SimpleTest.finish();
}
img.onload = function() {
  ok(false, "onload fired");
  SimpleTest.finish();
}
img.src = "";
</script>
</pre>
</body>
</html>
